flytt beregning av tilsagn til egne klasser
