+++ mulighetsrommet-api/src/main/kotlin/no/nav/mulighetsrommet/api/tilsagn/model/TilsagnBeregningForhandsgodkjent.kt
@@ -3,7 +3,10 @@ package no.nav.mulighetsrommet.api.tilsagn.model
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import no.nav.mulighetsrommet.domain.serializers.LocalDateSerializer
+import java.lang.Math.addExact
+import java.math.RoundingMode
 import java.time.LocalDate
+import kotlin.streams.asSequence
 
 @Serializable
 @SerialName("FORHANDSGODKJENT")
@@ -28,4 +31,29 @@ data class TilsagnBeregningForhandsgodkjent(
     data class Output(
         override val belop: Int,
     ) : TilsagnBeregningOutput()
+
+    companion object {
+        fun beregn(input: Input): TilsagnBeregningForhandsgodkjent {
+            val (periodeStart, periodeSlutt, sats, antallPlasser) = input
+
+            val output = periodeStart.datesUntil(periodeSlutt.plusDays(1))
+                .asSequence()
+                .groupBy { it.month }
+                .map { (_, datesInMonth) ->
+                    val fractionOfMonth = datesInMonth.size.toBigDecimal()
+                        .divide(datesInMonth[0].lengthOfMonth().toBigDecimal(), 2, RoundingMode.HALF_UP)
+
+                    val value = fractionOfMonth
+                        .multiply(sats.toBigDecimal())
+                        .multiply(antallPlasser.toBigDecimal())
+                        .setScale(0, RoundingMode.HALF_EVEN)
+
+                    value.intValueExact()
+                }
+                .reduce { acc: Int, s: Int -> addExact(acc, s) }
+                .let { Output(it) }
+
+            return TilsagnBeregningForhandsgodkjent(input, output)
+        }
+    }
 }
